+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Breaking** Removed `utils::copy_data`. It forked into a `/usr/bin/env cat`
   for copying. All internal uses of the function have been changed to simply
   use `std::io::copy` instead.
+- Replaced `nix` with `rustix`, following `wayland-rs`.
 
 ## v0.8.1 (7th Mar 2024)